requirements.txt
llama-index
llama-index-core
llama-index-llms-openai
llama-index-llms-groq
llama-index-experimental
python-dotenv
pandas
nltk